Add solution for two easy daily problems
Adds solution for check if two string arrays are equivalent easy daily
problem for 12/1/23 with time complexity of O(n) and space complexity
of O(1).
Adds solution for find words that can be formed by characters easy
daily problem for 12/2/23 with time complexity of O(n) and space
complexity of O(1).